rename of template file name
